+++ foundation_buffer/init.lua
@@ -1,4 +1,4 @@
-local mod = foundation.new_module("foundation_buffer", "2.0.0")
+local mod = foundation.new_module("foundation_buffer", "2.1.0")
 
 mod:require("string_buffer.lua")
 mod:require("token_buffer.lua")